+++ Arbiter.Net/Client/Messages/ClientRequestExitMessage.cs
@@ -12,8 +12,11 @@ public class ClientRequestExitMessage : ClientMessage
     public override void Deserialize(INetworkPacketReader reader)
     {
         base.Deserialize(reader);
-        
-        Reason = (ClientExitReason)reader.ReadByte();
+
+        if (reader.CanRead(1))
+        {
+            Reason = (ClientExitReason)reader.ReadByte();
+        }
     }
 
     public override void Serialize(INetworkPacketBuilder builder)